+++ demo/src/Form.js
@@ -24,7 +24,42 @@ const style = {
   },
 };
 
-// Our tiny component
+const unavailableUsernames = [
+  'elonmusk',
+  'ironman',
+  'lukeskywalker',
+];
+
+function validateLogin(value = '') {
+  if (value.trim() === '') {
+    return Promise.reject('Please enter an username');
+  }
+
+  return unavailableUsernames.includes(value)
+    ? Promise.reject('This username is unavailable')
+    : Promise.resolve()
+}
+
+function validatePassword(value = '') {
+  if (value.trim() === '') {
+    return Promise.reject('Please enter a password');
+  } else if (value.trim().length < 6) {
+    return Promise.reject('Password must contain 6 characters or more');
+  } else {
+    return Promise.resolve();
+  }
+}
+
+function validatePasswordConfirmation(value = '', password = '') {
+  if (value.trim() === '') {
+    return Promise.reject('Please enter a password');
+  } else if (value !== password) {
+    return Promise.reject('Please enter the same password as below');
+  } else {
+    return Promise.resolve();
+  }
+}
+
 export function Form({
   onSubmit,
 }) {
@@ -34,7 +69,8 @@ export function Form({
         name="login"
         component={Input}
         props={{
-          label: 'Login',
+          label: 'Login *',
+          onBlur: value => validateLogin(value),
           placeholder: 'elonmusk',
           type: 'string',
         }}
@@ -43,42 +79,55 @@ export function Form({
         name="pwd"
         component={Input}
         props={{
-          label: 'Password',
+          label: 'Password *',
+          onBlur: value => validatePassword(value),
           type: 'password',
         }}
       />
+      <Field
+        name="confirmPwd"
+        component={Input}
+        props={{
+          label: 'Password confirmation *',
+          onBlur: (value, { pwd }) => validatePasswordConfirmation(value, pwd),
+          type: 'password',
+        }}
+      />
+      <Field
+        name="referrer"
+        component={Input}
+        props={{
+          label: 'How did you find us? *',
+          placeholder: 'Google Search, Facebook or else',
+          type: 'string',
+        }}
+      />
       <button style={style.submitButton} type="submit">
-        Sign in
+        Sign up
       </button>
     </form>
   );
 }
 
-// Here's how we use hocForm
 export default hocForm({
   validate(values, props) {
     let errors = {};
+    const errorCatcher = (key, callback, ...args) => (
+      callback(values[key], args)
+        .catch(error => ({ [key]: error }))
+    );
 
-    if (!values.login) {
-      errors = {
-        ...errors,
-        login: 'Please enter a login',
-      };
-    }
+    return Promise.all([
+      errorCatcher('login', validateLogin),
+      errorCatcher('pwd', validatePassword),
+      errorCatcher('confirmPwd', validatePasswordConfirmation, values.pwd),
+    ]).then((errors) => {
+      if (!values.referrer) {
+        errors = errors.concat({ referrer: 'Please give us something! 😇🙏' });
+      }
 
-    if (!values.pwd) {
-      errors = {
-        ...errors,
-        pwd: 'Please enter a password',
-      };
-    }
-  
-    return errors;
-  },
-  asyncValidate(values) {
-    return values.pwd === 'starman'
-      ? Promise.resolve({})
-      : Promise.reject({ pwd: 'Invalid password' });
-  },
-  validateOnBlur: true,
+      const results = errors.reduce((acc, item) => ({ ...acc, ...item }), {});
+      return Object.keys(results).length ? Promise.reject(results) : Promise.resolve();
+    });
+  }
 })(Form);